[ YC-423 ] Add new property to ManagerEntity to allow selection of Shop's manager has access to

+++ trunk/manager/server/src/main/java/org/yes/cart/remote/service/impl/RemoteManagementServiceImpl.java
@@ -18,14 +18,19 @@ package org.yes.cart.remote.service.impl;
 
 import org.yes.cart.domain.dto.ManagerDTO;
 import org.yes.cart.domain.dto.RoleDTO;
+import org.yes.cart.domain.dto.ShopDTO;
 import org.yes.cart.exception.UnableToCreateInstanceException;
 import org.yes.cart.exception.UnmappedInterfaceException;
 import org.yes.cart.remote.service.RemoteManagementService;
 import org.yes.cart.service.dto.ManagementService;
+import org.yes.cart.service.federation.FederationFacade;
 
 import java.io.UnsupportedEncodingException;
 import java.security.NoSuchAlgorithmException;
+import java.util.Collections;
+import java.util.Iterator;
 import java.util.List;
+import java.util.Set;
 
 /**
  * User: Igor Azarny iazarny@yahoo.com
@@ -35,23 +40,30 @@ import java.util.List;
 public class RemoteManagementServiceImpl implements RemoteManagementService {
 
     private final ManagementService managementService;
+    private final FederationFacade federationFacade;
 
     /**
-     * Construct remote user managment service.
+     * Construct remote user management service.
      *
      * @param managementService
+     * @param federationFacade
      */
-    public RemoteManagementServiceImpl(final ManagementService managementService) {
+    public RemoteManagementServiceImpl(final ManagementService managementService,
+                                       final FederationFacade federationFacade) {
         this.managementService = managementService;
+        this.federationFacade = federationFacade;
     }
 
 
     /**
      * {@inheritDoc}
      */
-    public void addUser(final String userId, final String firstName, final String lastName)
+    public void addUser(final String userId, final String firstName, final String lastName, final String shopCode)
             throws NoSuchAlgorithmException, UnsupportedEncodingException {
-        managementService.addUser(userId, firstName, lastName);
+
+        if (federationFacade.isShopAccessibleByCurrentManager(shopCode)) {
+            managementService.addUser(userId, firstName, lastName, shopCode);
+        }
     }
 
     /**
@@ -59,7 +71,10 @@ public class RemoteManagementServiceImpl implements RemoteManagementService {
      */
     public List<ManagerDTO> getManagers(final String emailFilter, final String firstNameFilter, final String lastNameFilter)
             throws UnmappedInterfaceException, UnableToCreateInstanceException {
-        return managementService.getManagers(emailFilter, firstNameFilter, lastNameFilter);
+
+        final List<ManagerDTO> managers = managementService.getManagers(emailFilter, firstNameFilter, lastNameFilter);
+        federationFacade.applyFederationFilter(managers, ManagerDTO.class);
+        return managers;
     }
 
     /**
@@ -67,7 +82,10 @@ public class RemoteManagementServiceImpl implements RemoteManagementService {
      */
     public List<RoleDTO> getAssignedManagerRoles(String userId)
             throws UnmappedInterfaceException, UnableToCreateInstanceException {
-        return managementService.getAssignedManagerRoles(userId);
+        if (federationFacade.isManageable(userId, ManagerDTO.class)) {
+            return managementService.getAssignedManagerRoles(userId);
+        }
+        return Collections.emptyList();
     }
 
     /**
@@ -75,28 +93,37 @@ public class RemoteManagementServiceImpl implements RemoteManagementService {
      */
     public List<RoleDTO> getAvailableManagerRoles(String userId)
             throws UnmappedInterfaceException, UnableToCreateInstanceException {
-        return managementService.getAvailableManagerRoles(userId);
+        if (federationFacade.isManageable(userId, ManagerDTO.class)) {
+            return managementService.getAvailableManagerRoles(userId);
+        }
+        return Collections.emptyList();
     }
 
     /**
      * {@inheritDoc}
      */
     public void updateUser(final String userId, final String firstName, final String lastName) {
-        managementService.updateUser(userId, firstName, lastName);
+        if (federationFacade.isManageable(userId, ManagerDTO.class)) {
+            managementService.updateUser(userId, firstName, lastName);
+        }
     }
 
     /**
      * {@inheritDoc}
      */
     public void resetPassword(final String userId) {
-        managementService.resetPassword(userId);
+        if (federationFacade.isManageable(userId, ManagerDTO.class)) {
+            managementService.resetPassword(userId);
+        }
     }
 
     /**
      * {@inheritDoc}
      */
     public void deleteUser(final String userId) {
-        managementService.deleteUser(userId);
+        if (federationFacade.isManageable(userId, ManagerDTO.class)) {
+            managementService.deleteUser(userId);
+        }
     }
 
 
@@ -112,8 +139,8 @@ public class RemoteManagementServiceImpl implements RemoteManagementService {
     /**
      * {@inheritDoc}
      */
-    public void addRole(final String role, final String decription) {
-        managementService.addRole(role, decription);
+    public void addRole(final String role, final String description) {
+        managementService.addRole(role, description);
     }
 
     /**
@@ -134,14 +161,91 @@ public class RemoteManagementServiceImpl implements RemoteManagementService {
      * {@inheritDoc}
      */
     public void grantRole(final String userId, final String role) {
-        managementService.grantRole(userId, role);
+        if ("ROLE_SMADMIN".equals(role) && !federationFacade.isCurrentUserSystemAdmin()) {
+            return;
+        }
+        if (federationFacade.isManageable(userId, ManagerDTO.class)) {
+            managementService.grantRole(userId, role);
+        }
     }
 
     /**
      * {@inheritDoc}
      */
     public void revokeRole(final String userId, final String role) {
-        managementService.revokeRole(userId, role);
+        if ("ROLE_SMADMIN".equals(role) && !federationFacade.isCurrentUserSystemAdmin()) {
+            return;
+        }
+        if (federationFacade.isManageable(userId, ManagerDTO.class)) {
+            managementService.revokeRole(userId, role);
+        }
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public List<ShopDTO> getAssignedManagerShops(final String userId) throws UnmappedInterfaceException, UnableToCreateInstanceException {
+        if (federationFacade.isManageable(userId, ManagerDTO.class)) {
+            final List<ShopDTO> assigned = managementService.getAssignedManagerShops(userId);
+
+            if (!federationFacade.isCurrentUserSystemAdmin()) { // restrict other managers
+                final Set<Long> currentAssignedIds = federationFacade.getAccessibleShopIdsByCurrentManager();
+                final Iterator<ShopDTO> availableIt = assigned.iterator();
+                while (availableIt.hasNext()) {
+                    final ShopDTO shop = availableIt.next();
+                    if (!currentAssignedIds.contains(shop.getShopId())) {
+                        availableIt.remove();
+                    }
+                }
+            }
+            return assigned;
+        }
+        return Collections.emptyList();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public List<ShopDTO> getAvailableManagerShops(final String userId) throws UnmappedInterfaceException, UnableToCreateInstanceException {
+
+        if (federationFacade.isManageable(userId, ManagerDTO.class)) {
+            final List<ShopDTO> available = managementService.getAvailableManagerShops(userId);
+
+            if (!federationFacade.isCurrentUserSystemAdmin()) { // restrict other managers
+                final Set<Long> currentAssignedIds = federationFacade.getAccessibleShopIdsByCurrentManager();
+                final Iterator<ShopDTO> availableIt = available.iterator();
+                while (availableIt.hasNext()) {
+                    final ShopDTO shop = availableIt.next();
+                    if (!currentAssignedIds.contains(shop.getShopId())) {
+                        availableIt.remove();
+                    }
+                }
+            }
+            return available;
+        }
+        return Collections.emptyList();
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public void grantShop(final String userId, final String shopCode) {
+
+        if (federationFacade.isManageable(userId, ManagerDTO.class)
+                && federationFacade.isShopAccessibleByCurrentManager(shopCode)) {
+            managementService.grantShop(userId, shopCode);
+        }
+
+    }
+
+    /**
+     * {@inheritDoc}
+     */
+    public void revokeShop(final String userId, final String shopCode) {
+        if (federationFacade.isManageable(userId, ManagerDTO.class)
+                && federationFacade.isShopAccessibleByCurrentManager(shopCode)) {
+            managementService.revokeShop(userId, shopCode);
+        }
     }
 
 }
